Fix ragged bottom row in scrolling grid widgets
Puts grid thumbs into a container and sets its height to the height of the shortest column after render. Some of the bottom row may be obscured when the widget is scrolled all the way down, but that's much better-looking than the possibility of one very tall thumb pushing everything else up and out of the viewport.