Override cascading transforms on images
Embedding pinterest widgets is not compatible with the default theme of Ghost, which applies a default transition to images inside the post content. This causes all the images to appear pulled 50% to the left. (See example: http://puu.sh/knIVj.png)

There is no top-level class or id on the widget to hook into, which makes it very tricky to restyle, so I've submitted a fix - setting images to `transform: inherit` ensures that the widget images are no longer affected. I'd also recommend having a top-level class that can be used to fix any other such issues.