push new buttons and widgets to 100%
A pretty complete rewrite of pinit_main.js, which we shipped to 100% a week ago. 

We've done our best not to break existing installs and to respect existing social conventions like data-pin-do=none, nothing, and ignore. I'm going to close out old issues and PRs with this push; very sorry for the radio silence but it's been a busy time.